backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
websockets==12.0
python-dotenv
pydantic==2.11.1
cua-computer==0.3.0
cua-agent==0.2.15
python-multipart==0.0.6
aiofiles==23.2.1 